Adds dynamic interval support.
Adds a `-f` parameter for changing the frequency of the pings. If passed
value is not an integer or is below 1, it is ignored in favor of the
default.

Consideration should be given to the fact that changing the frequency impacts
both the time to see the first output and the wrap-up of the script when killed
via Ctrl+c.